Configure a default set of KSM custom resources (#280)
* Add support for defining pre-configured KSM custom resources

With this, we can configure extra KSM custom resources without the need to specify the complete resource configuration.

Towards https://github.com/giantswarm/giantswarm/issues/32502

* Simplify configuration for the KSM custom resources

With this, only a simple string is needed to enable a resource. All configuration is abstracted away.

* Docs

* Use an object instead of a list

This way the value can be merged from multiple config sources.

* Use $

* Set default ksm CRs if cluster is an MC

If the cluster is an MC, the chart will pre-configure multiple custom resources in KSM, based on the CAPI provider.

* Configure some additional KSM CRs

* Default to empty map

* Always enable all KSM custom resources

KSM builds a map of all available CRDs, and just silently ignores all those configured custom resources that are not present in the cluster. So we can always enable all of them, KSM will only generate the metrics of those available

* Finish defining all CRs that are in the capi-monitoring app

* Cleanup

* Change the order of the generated ksm resources list

When two items in the `resources` list share the same api group, version and kind, KSM only uses the last occurrence, discarding the previous ones. So we put the default generated resources at the beginning of the list, to give priority to the user-defined resources